+++ lib/src/analyzers/lint_analyzer/rules/rules_list/format_comment/models/comment_type.dart
@@ -1,3 +1,0 @@
-part of '../format_comment_rule.dart';
-
-enum CommentType { base, multiline, documentation }