+++ lib/src/analyzers/lint_analyzer/rules/rules_list/format_single_line_comment/models/comment_type.dart
@@ -1,0 +1,3 @@
+part of '../format_single_line_comment_rule.dart';
+
+enum CommentType { base, documentation }