+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* feat: add static code diagnostics `format-comment`.
+* feat: add static code diagnostics `format-single-line-comment`.
 * fix: ignore a class usage inside `State<T>` for `check-unused-code` command.
 * fix: correctly handle variables declaration for `check-unused-code` command.
 * feat: add static code diagnostics `avoid-dynamic`, `prefer-async-await`.